backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
google-cloud-vision==2.4.2
python-dotenv==0.19.0
requests==2.26.0
